telegram-bot/requirements.txt
python-telegram-bot
requests>=2.31.0
bonbast>=1.0.0
aiohttp>=3.8.0
